fix: 수입 화면 디자인 변경

+++ core/domain/src/main/java/jun/money/mate/domain/AddIncomeUsecase.kt
@@ -14,7 +14,7 @@ class AddIncomeUsecase @Inject constructor(
     suspend operator fun invoke(
         id: Long,
         title: String,
-        amount: Double,
+        amount: Long,
         type: IncomeType,
         incomeDate: LocalDate,
         onSuccess: () -> Unit,
